# Conflicts:
#	src/main/java/com/example/finanzasgrupo5backend/Credito2/Repository/Consumo/IConsumoCredito2Repository.java

+++ src/main/java/com/example/finanzasgrupo5backend/Formulas/Anualidades/SimpleVencida/AnualidadSimpleVencida.java
@@ -111,7 +111,7 @@ public class AnualidadSimpleVencida {
         return renta;
     }
 
-    public static Double calcularInteresCompensatorio(String tep_credito, Double tasa_credito, Double valorNominal, long diasAtraso) {
+    public static Double calcularInteresCompensatorio(String tep_credito, Double tasa_credito, Double valorNominal, Double diasAtraso) {
 
         //Ic= valor nominal [(1+tep)^(dias trasladar/ dias periodo)  -1]
         Double ndp = numeroDiasPeriodo(tep_credito);
@@ -122,7 +122,7 @@ public class AnualidadSimpleVencida {
         return interesCompensatorio;
     }
 
-    public static Double calcularInteresMoratorio(String tep_mora, Double tasa_mora, Double valorNominal, long diasAtraso) {
+    public static Double calcularInteresMoratorio(String tep_mora, Double tasa_mora, Double valorNominal, Double diasAtraso) {
 
         //Im= valor nominal [(1+tep)^(dias trasladar/ dias periodo)  -1]
         Double ndp = numeroDiasPeriodo(tep_mora);
